+++ example/websocket/client/async/websocket_client_async.cpp
@@ -95,7 +95,7 @@ public:
     }
 
     void
-    on_connect(beast::error_code ec, tcp::resolver::results_type::endpoint_type)
+    on_connect(beast::error_code ec, tcp::resolver::results_type::endpoint_type ep)
     {
         if(ec)
             return fail(ec, "connect");
@@ -117,6 +117,11 @@ public:
                     std::string(BOOST_BEAST_VERSION_STRING) +
                         " websocket-client-async");
             }));
+
+        // Update the host_ string. This will provide the value of the
+        // Host HTTP header during the WebSocket handshake.
+        // See https://tools.ietf.org/html/rfc7230#section-5.4
+        host_ += ':' + std::to_string(ep.port());
 
         // Perform the websocket handshake
         ws_.async_handshake(host_, "/",